image-classifier/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
tensorflow==2.17.*
numpy
pytest
